pyproject.toml
[build-system]
requires = ["setuptools>=42.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "quickscrape"
version = "0.1.0"
description = "A flexible and powerful web scraping tool"
authors = [
    {name = "QuickScrape Team", email = "info@quickscrape.io"}
]
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Internet :: WWW/HTTP :: Browsers",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "click>=8.1.3",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "beautifulsoup4>=4.11.1",
    "requests>=2.28.1",
    "cloudscraper>=1.2.60",
    "playwright>=1.39.0",
    "pandas>=2.0.0",
    "openpyxl>=3.0.10",
    "nest-asyncio>=1.5.8",
    "httpx>=0.24.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-mock>=3.10.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.23.0",
    "mypy>=1.0.0",
    "ruff>=0.0.261",
    "black>=23.3.0",
]

[project.scripts]
quickscrape = "quickscrape.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.ruff]
line-length = 88
target-version = "py310"
select = ["E", "F", "W", "I", "N", "UP", "ANN", "B", "COM", "C4", "T10", "RET", "SIM", "ERA"]
ignore = ["ANN101", "ANN102"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
python_classes = "Test*"
addopts = "--cov=quickscrape --cov-report=term-missing"
markers = [
    "unit: marks tests as unit tests",
    "integration: marks tests as integration tests",
    "asyncio: mark tests as asyncio tests",
]

[tool.mypy]
python_version = "3.10"
warn_redundant_casts = true
warn_unused_ignores = true
disallow_any_generics = true
check_untyped_defs = true
no_implicit_reexport = true
disallow_untyped_defs = true 